# docs/contracts/http/commands.openapi.yaml
openapi: 3.1.0
info:
  title: HomeTusk Commands API
  version: 1.0.0
  description: |
    Intent-driven API for household task management.
    Users submit commands, not CRUD operations.

    **Stage 1:** Structured commands (no NL processing yet).

    See also:
    - [CLAUDE.md](../../../CLAUDE.md) — Development rules
    - [ADR-003](../../architecture/decisions/003-stage1-commands-api.md) — Stage 1 decisions
servers:
  - url: /api/v1
    description: API version 1
security:
  - bearerAuth: []
...
paths:
...
  /households/{householdId}/members:
    get:
      operationId: listHouseholdMembers
      summary: List members of a household
      tags:
        - Households
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HouseholdMember'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Not a member of this household
  /households/{householdId}/zones:
    get:
      operationId: listZones
      summary: List zones in a household
      tags:
        - Households
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Zone'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Not a member of this household
    post:
      operationId: createZone
      summary: Create a zone in household
      tags:
        - Households
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateZoneRequest'
      responses:
        '201':
          description: Zone created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Not a member of this household
  /households/{householdId}/invites:
    post:
      operationId: createHouseholdInvite
      summary: Create a household invite
      tags:
        - Invites
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInviteResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Not a member of this household
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
  /invites/accept:
    post:
      operationId: acceptHouseholdInvite
      summary: Accept a household invite
      tags:
        - Invites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInviteRequest'
      responses:
        '200':
          description: Invite accepted or no-op for existing member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptInviteResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '404':
          description: Invalid invite token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
        '410':
          description: Invite expired, redeemed, or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
...
  /households:
    post:
      operationId: createHouseholdPublic
      summary: Create a new household
      description: |
        Creates a new household and automatically adds the current user as admin.
        This is the user-facing endpoint for household creation.
      tags:
        - Households
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseholdRequest'
      responses:
        '201':
          description: Household created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '400':
          description: Invalid request (name validation failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
        '401':
          description: Authentication required
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak JWT token
  schemas:
...
    CommandErrorResponse:
      type: object
      required:
        - correlationId
        - errorCode
        - message
      properties:
        correlationId:
          type: string
          format: uuid
        errorCode:
          type: string
          description: Machine-readable error code (see ErrorCode enum)
        message:
          type: string
          description: Human-readable error message
        validationErrors:
          type: array
          description: Present when errorCode is SCHEMA_INVALID
          items:
            $ref: '#/components/schemas/ValidationError'
        violations:
          type: array
          description: Present when errorCode is BUSINESS_RULE_VIOLATION
          items:
            $ref: '#/components/schemas/BusinessViolation'
    ValidationError:
      type: object
      required:
        - path
        - code
        - message
      properties:
        path:
          type: string
          description: JSON path to invalid field (e.g., $.payload.deadline)
        code:
          type: string
          description: Validation error code
        message:
          type: string
          description: Human-readable validation message
    BusinessViolation:
      type: object
      required:
        - rule
        - message
      properties:
        rule:
          type: string
          enum:
            - ASSIGNEE_MUST_BE_MEMBER
            - ZONE_MUST_EXIST
            - DEADLINE_MUST_BE_FUTURE
            - INITIATOR_MUST_HAVE_PERMISSION
            - TASK_ALREADY_COMPLETED
          description: Business rule that was violated
        message:
          type: string
          description: Human-readable explanation
    AuthErrorResponse:
      type: object
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - AUTH_TOKEN_MISSING
            - AUTH_TOKEN_INVALID
            - AUTH_TOKEN_EXPIRED
        message:
          type: string

    # Domain entities
...
    Household:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    HouseholdMember:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, member]
        joinedAt:
          type: string
          format: date-time
    Zone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        householdId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    CreateHouseholdRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
          description: Household name (trimmed, non-blank)
    CreateZoneRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
    AcceptInviteRequest:
      type: object
      required:
        - inviteToken
      properties:
        inviteToken:
          type: string
          description: Invite token (hti_ prefix)
    CreateInviteResponse:
      type: object
      required:
        - inviteToken
        - expiresAt
        - status
      properties:
        inviteToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, redeemed, expired, revoked]
        inviteLink:
          type: string
          nullable: true
    InviteMembership:
      type: object
      required:
        - id
        - role
        - joinedAt
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [admin, member]
        joinedAt:
          type: string
          format: date-time
    AcceptInviteResponse:
      type: object
      required:
        - membership
        - household
      properties:
        membership:
          $ref: '#/components/schemas/InviteMembership'
        household:
          $ref: '#/components/schemas/Household'

    # Notification schemas
...